safari-600.7-branch/LayoutTests/editing/shadow/contenteditable-propagation-at-shadow-boundary-expected.txt

This test checks that content-editable is not propagated from shadow host to a shadow subtree.

This p is required to produce the issue.

This div and parent div are required to produce the issue.
PASS computedStyle(nodeInShadowRoot1, userModifyPropertyName) is "read-only"
PASS successfullyParsed is true

TEST COMPLETE
